Keypad MISRA V1

+++ HAL/Keypad_OS.h
@@ -3,7 +3,8 @@
 
 #define NO_OF_ROWS ((u8)4)
 #define NO_OF_COLS ((u8)3)
-#define KEYPAD_TASK_PERIODICTIY 200
+
+#define KEYPAD_TASK_PERIODICTIY ((u32)200)
 
 void KEYPAD_OS_vidInit(void);
 void KEYPAD_OS_Task(void *pvoid);
